Turn indexation into a proper module
